requirements.txt
absl-py==0.7.1
astor==0.8.0
blessings==1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
Cython==0.29.12
future==0.17.1
gast==0.2.2
glfw==1.8.2
google-pasta==0.1.7
gpustat==0.5.0
grpcio==1.22.0
gym==0.13.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.1.1
matplotlib==3.1.1
mujoco-py==2.0.2.4
numpy==1.16.4
nvidia-ml-py3==7.352.0
# opencv-python==4.1.0.25
pillow>=6.2.2
protobuf==3.8.0
psutil==5.6.6
pycparser==2.19
pyglet==1.3.2
pyparsing==2.4.0
python-dateutil==2.8.0
pyzmq==18.0.2
requests==2.22.0
# scipy==1.3.0
setuptools==41.0.1
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
# tensorflow-gpu==1.15.2
termcolor==1.1.0
torch==1.10.0
torchfile==0.1.0
# torchvision==0.3.0
tornado==6.0.3
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.2
xmltodict==0.12.0
wheel==0.33.6
baselines==0.1.5
git+git://github.com/CampagneLaboratory/torchfold@c2ac41b#egg=torchfold

